fixed major bug with ftp connection being open often and for no reason
Essentially, in flush_writes and flush_deletes methods, if there is nothing to write or delete
(frequent case), ftp connection was being open just to immediately close it. Silly, and slowed down
all aspects of the system, not just the image uploading part.

Fix is simple - do not bother opening ftp connection if only goal is to close it.