removed dependency on paperclip
Paperclip is one of the most forked projects, and many people are making their own versions of
paperclip gem. Hard linking a dependency on paperclip here is a disservice to them.

It is expected that people that want to use paperclipftp plugin already know
about paperclip gem and already added it to their environment.